# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for pickup stats collected with LogWatcher.",
        "title": "Pickup Stats API",
        "contact": {}
    },
    "basePath": "/api",
    "paths": {
        "/dpm": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ratings"
                ],
                "summary": "Player rating by average DPM.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player class",
                        "name": "class",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum games played",
                        "name": "mingames",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gamesCount": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Util"
                ],
                "summary": "Games count in mongodb.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GamesCount"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hpm": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ratings"
                ],
                "summary": "Medics rating by average heals given per minute.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Minimum games played",
                        "name": "mingames",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/kdr": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ratings"
                ],
                "summary": "Player rating by average KDR.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player class",
                        "name": "class",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum games played",
                        "name": "mingames",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.GamesCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Result"
                    }
                }
            }
        },
        "db.Result": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "dpm": {
                    "type": "number"
                },
                "games": {
                    "type": "integer"
                },
                "hpm": {
                    "type": "number"
                },
                "kdr": {
                    "type": "number"
                },
                "player_name": {
                    "type": "string"
                },
                "steamid64": {
                    "type": "string"
                }
            }
        }
    }
}